Refactor movie search to use options object
This commit refactors the movie search functionality to use an options object instead of individual parameters. The `GetAllMoviesRequest` class is added to the contracts project, and a corresponding `GetAllMoviesOptions` class is added to the application project. A new validator for `GetAllMoviesOptions` is also added. The `MoviesController.GetAll()` method now accepts a `GetAllMoviesRequest` parameter, which is mapped to an instance of `GetAllMoviesOptions`. The changes allow for more flexibility in searching movies by title, year of release, and user ID.