Refactor GetAllMoviesOptions to include sorting options
This commit adds the ability to sort movies by title or year of release. The new SortField and SortOrder properties have been added to the GetAllMoviesOptions class, which can be used in conjunction with the existing Title and YearOfRelease properties. The MovieRepository has also been updated to include an ORDER BY clause based on these new options. Additionally, a validation rule has been added to ensure that only acceptable fields are used for sorting.